+++ src/app/admin/admin.module.ts
@@ -33,7 +33,6 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { NgxMatSelectSearchModule } from 'ngx-mat-select-search';
 
-import { TranslocoRootModule } from '@app/transloco-root.module';
 import { TRANSLOCO_SCOPE } from '@jsverse/transloco';
 
 import { BlockUIModule } from 'ng-block-ui';
@@ -138,75 +137,74 @@ const material = [
 ];
 
 @NgModule({
-    imports: [
-        BlockUIModule.forRoot({
-            delayStart: 500,
-            /*delayStop: 500*/
-        }),
-        CommonModule,
-        AdminRoutingModule,
-        material,
-        FormsModule,
-        ReactiveFormsModule,
-        SharedModule,
-        TranslocoRootModule,
-        // Directives
-        DisableControlDirective,
-        // Components
-        AdminComponent,
-        SideMenuComponent,
-        OrthographyComponent,
-        NavMenuComponent,
-        NavItemComponent,
-        UiLanguageComponent,
-        UiScopeComponent,
-        UiTranslationComponent,
-        UiResultCategoryComponent,
-        LanguageComponent,
-        LangPairComponent,
-        LangOrthoMappingComponent,
-        LexemeTypeComponent,
-        LexemeFormTypeComponent,
-        UnitLevelComponent,
-        CategoryComponent,
-        ViewComponent,
-        ListComponent,
-        EditorComponent,
-        LexemeLinkComponent,
-        TabGeneralComponent,
-        TabVariantsComponent,
-        TabSememesComponent,
-        TabLinksComponent,
-        TabCommentsComponent,
-        OtherTranslationsComponent,
-        DisableFormComponent,
-        LexemeFormComponent,
-        LemmaTemplateComponent,
-        TypeLanguageConfigComponent,
-        FormTypePositionsDialog,
-        LexemeFormsComponent,
-        YesNoDialogComponent,
-        FilterMenuComponent,
-        ErrorsDialogComponent,
-        LinkTypeComponent,
-        TabMappingsComponent,
-        TabHistoryComponent,
-        SynGroupLinkComponent,
-        SememeLinkComponent,
-        HierarchicalSelectComponent,
-        DialectsSelectComponent
-    ],
-    providers: [
-        httpInterceptorProviders,
-        { provide: TRANSLOCO_SCOPE, useValue: 'admin' },
-        DataService,
-        LexemeQueryService,
-        LexemeService,
-        SememeService,
-        SynGroupQueryService,
-        LockService,
-        NavigationService,
-        provideHttpClient(withInterceptorsFromDi()),
-    ],
+	imports: [
+		BlockUIModule.forRoot({
+			delayStart: 500,
+			//delayStop: 500,
+		}),
+		CommonModule,
+		AdminRoutingModule,
+		material,
+		FormsModule,
+		ReactiveFormsModule,
+		SharedModule,
+		// Directives
+		DisableControlDirective,
+		// Components
+		AdminComponent,
+		SideMenuComponent,
+		OrthographyComponent,
+		NavMenuComponent,
+		NavItemComponent,
+		UiLanguageComponent,
+		UiScopeComponent,
+		UiTranslationComponent,
+		UiResultCategoryComponent,
+		LanguageComponent,
+		LangPairComponent,
+		LangOrthoMappingComponent,
+		LexemeTypeComponent,
+		LexemeFormTypeComponent,
+		UnitLevelComponent,
+		CategoryComponent,
+		ViewComponent,
+		ListComponent,
+		EditorComponent,
+		LexemeLinkComponent,
+		TabGeneralComponent,
+		TabVariantsComponent,
+		TabSememesComponent,
+		TabLinksComponent,
+		TabCommentsComponent,
+		OtherTranslationsComponent,
+		DisableFormComponent,
+		LexemeFormComponent,
+		LemmaTemplateComponent,
+		TypeLanguageConfigComponent,
+		FormTypePositionsDialog,
+		LexemeFormsComponent,
+		YesNoDialogComponent,
+		FilterMenuComponent,
+		ErrorsDialogComponent,
+		LinkTypeComponent,
+		TabMappingsComponent,
+		TabHistoryComponent,
+		SynGroupLinkComponent,
+		SememeLinkComponent,
+		HierarchicalSelectComponent,
+		DialectsSelectComponent
+	],
+	providers: [
+		httpInterceptorProviders,
+		{ provide: TRANSLOCO_SCOPE, useValue: 'admin' },
+		DataService,
+		LexemeQueryService,
+		LexemeService,
+		SememeService,
+		SynGroupQueryService,
+		LockService,
+		NavigationService,
+		provideHttpClient(withInterceptorsFromDi()),
+	],
 })
 export class AdminModule { }
